Break down order schemas and use DI
